fix(has_Support): check that two rows/cols are not nonzero in only one shared col/row respectively
Previous check for whether the matrix has total support only checked for rows and cols with all zeros in them (by taking a dot product with a vector of 1s). However, this check misses the case where some pair of rows/cols have all zeros except for in one shared location, e.g. two rows have zeros everywhere but in their first element. In this case, we'd be unable to permute the matrix in such a way as to fulfill the condition of creating a positive diagonal out of those two problematic rows/cols.